Move Gen to its own source file
Also rename `PropertyBasedTest` to `PropertyTest`.
